Rename createADir endpoint as createDir
